防止SQL注入过滤

+++ python/app/model/Model.py
@@ -2,6 +2,7 @@ from app.Env import Env
 from app.common.Base import Base
 from dbutils.pooled_db import PooledDB
 import pymysql
+import re
 
 # 数据库
 class Model(Base) :
@@ -9,6 +10,7 @@ class Model(Base) :
   _conn = None  #链接
   _state = False #事务
   _sqlAll = [] #记录SQL
+  _sql_reg = r"(?:')|(?:--)|(/\*(?:.|[\n\r])*?\*/)|(\b(select|select|update|union|and|or|delete|insert|trancate|char|into|substr|ascii|declare|exec|count|master|into|drop|execute)\b)"
 
   # 链接
   def __init__(self) :
@@ -16,15 +18,32 @@ class Model(Base) :
       pool = PooledDB(**Env.db())
       self._conn = pool.connection()
 
+  # 过滤-参数值
+  def filter(self, str=''):
+    return re.sub(r'.*([\';]+|(--)+).*','',str)
+  # 过滤-WHERE
+  def bindWhere(self,where,bind):
+    for k in bind.keys() :
+      v = str(bind[k])
+      # 小写、匹配、替换
+      lower = v.lower()
+      if re.search(self._sql_reg,lower) :
+        print('SQL过滤: '+v)
+        return ''
+      where = re.sub(r':'+k+':', v, where)
+    return where
+
   # 查询-单条
   def findFirst(self,params={}):
-    sql = self.sql('SELECT',params)
+    sql = self._getSql('SELECT',params)
+    if sql=='' : return None
     data = self.fetchOne(sql)
     return data
 
   # 查询-多条
   def find(self,params={}):
-    sql = self.sql('SELECT',params)
+    sql = self._getSql('SELECT',params)
+    if sql=='' : return None
     data = self.fetchAll(sql)
     return data
 
@@ -33,7 +52,7 @@ class Model(Base) :
     keys = params.keys()
     vals = ''
     for name in keys :
-      vals += '%s,'%(params[name]) if params[name]=='null' else '"%s",'%(params[name])
+      vals += '%s,'%params[name] if params[name]=='null' else '"%s",'%self.filter(params[name])
     sql = 'INSERT INTO %s(`%s`) values(%s)'%(self.table,'`,`'.join(keys),vals[:-1])
     # 提交
     if self._state :
@@ -53,11 +72,24 @@ class Model(Base) :
       
   # 更新
   def update(self,params,where=''):
-    # SQL
-    keys = params.keys()
+    # 是否数据
+    if 'data' not in params.keys():
+      print('请传入更新数据!')
+      return False
+    # 是否WHERE
+    if 'where' not in params.keys() or params['where']=='' :
+      print('必需传入Where条件!')
+      return False
+    # 过滤WHERE
+    where = params['where']
+    if 'bind' in params.keys():
+      where = self.bindWhere(where,params['bind'])
+      if where=='' : return False
+    # 组合SQL
+    keys = params['data'].keys()
     vals = ''
-    for name in keys:
-      vals += '%s="%s",'%(name,params[name])
+    for k in keys:
+      vals += '%s="%s",'%(k,params['data'][k])
     sql = 'UPDATE `%s` SET %s WHERE %s'%(self.table,vals[:-1],where)
     # 提交
     if self._state :
@@ -76,8 +108,18 @@ class Model(Base) :
       return res
 
   # 删除
-  def delete(self,where=''):
-    # SQL
+  def delete(self,params={}):
+    keys = params.keys()
+    # 是否WHERE
+    if 'where' not in keys or params['where']=='' :
+      print('必需传入Where条件!')
+      return False
+    # 过滤WHERE
+    where = params['where']
+    if 'bind' in keys :
+      where = self.bindWhere(where,params['bind'])
+      if where=='' : return False
+    # 组合SQL
     sql = 'DELETE FROM `%s` WHERE %s'%(self.table,where)
     # 提交
     if self._state :
@@ -96,16 +138,25 @@ class Model(Base) :
       return res
 
   # 组合SQL
-  def sql(self,type,params={}):
+  def _getSql(self,type,params={}):
+    keys = params.keys()
     # 参数
-    columns = params['columns'] if 'columns' in params else '*'
-    table = params['table'] if 'table' in params else self.table
+    columns = params['columns'] if 'columns' in keys else '*'
+    table = params['table'] if 'table' in keys else self.table
     # SQL
     sql = '%s %s FROM %s'%(type,columns,table)
-    if 'where' in params : sql += ' WHERE %s'%params['where']
-    if 'group' in params : sql += ' GROUP BY %s'%params['group']
-    if 'order' in params : sql += ' ORDER BY %s'%params['order']
-    if 'limit' in params : sql += ' LIMIT %s'%params['limit']
+    # Where
+    if 'where' in keys :
+      # 过滤WHERE
+      where = params['where']
+      if 'bind' in keys :
+        where = self.bindWhere(where,params['bind'])
+        if where=='' : return ''
+      sql += ' WHERE '+where
+    # Group、Order、limit
+    if 'group' in params : sql += ' GROUP BY '+params['group']
+    if 'order' in params : sql += ' ORDER BY '+params['order']
+    if 'limit' in params : sql += ' LIMIT '+params['limit']
     return sql
 
   # 查询-单条
